chatbot/requirements.txt
flask
faiss-cpu
sentence-transformers
pytesseract
pymupdf
pillow
google-generativeai
python-dotenv
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-ollama
langchain-google-genai